Bug correction x2

+++ src/ExcersiceBalance/StackExercise.java
@@ -47,7 +47,7 @@ public class StackExercise {
 	public static void main(String[] args) throws NumberFormatException, IOException {
 		long index = Long.parseLong(bf.readLine());
 		String line = bf.readLine();
-		for (int i = 0; i < index; i++) {
+		for (long i = 0; i < index; i++) {
 			System.out.println(balance(line));
 			line = bf.readLine();
 		}
